Optimize JSX import source detection
