+++ .changeset/wicked-dolphins-design.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Optimize JSX import source detection
